Extract identity persistence orchestration
