+++ Sources/ClerkKit/Core/Clerk.swift
@@ -35,47 +35,6 @@ public final class Clerk {
   private static var isRuntimeReconfigurationInProgress = false
   private static var runtimeReconfigurationWaiters: [CheckedContinuation<Void, Never>] = []
 
-  private struct ReconfigurationRollbackState {
-    let configurationEpoch: ClerkConfigurationEpoch
-    let dependencies: any Dependencies
-    let identity: ClerkIdentityController.RollbackState
-  }
-
-  private struct PreparedReconfiguration {
-    let ownership: PersistenceTransitionOwnership
-    let nextEpoch: ClerkConfigurationEpoch
-    let dependencies: DependencyContainer
-    let plan: ReconfigurationPlan
-    let rollbackState: ReconfigurationRollbackState
-    let volatileIdentity: ClerkIdentitySnapshot?
-  }
-
-  private enum AppContainerIdentityClearBootstrapResult {
-    case notNeeded
-    case recovered
-    case blocked(
-      reason: PersistenceBlockReason,
-      recoveryFailure: PersistenceFailureKind,
-      requiresAtomicRouting: Bool
-    )
-  }
-
-  typealias ConfigurationDependencyFactory = @MainActor (
-    String,
-    Clerk.Options,
-    PersistenceFailureKind?
-  ) throws -> DependencyContainer
-
-  private enum ReconfigurationPlan: Equatable {
-    case preserveIdentityAndSlot
-    case preserveIdentityAndMigrateSlot
-    case replaceIdentity
-
-    var reusesOwnerSlot: Bool {
-      self == .preserveIdentityAndSlot
-    }
-  }
-
   /// A getter to see if the Clerk object is ready for use or not.
   /// Returns true when persisted identity is safe to use and both environment
   /// and client are loaded.
@@ -303,21 +262,6 @@ public final class Clerk {
   /// concurrent sibling changes without assuming either side always wins.
   var sharedSessionDegradationBaseGeneration: UInt64?
 
-  var sharedSessionLocalMutation: SharedSessionLocalMutation? {
-    guard case .durable =
-      identityPersistenceOperationCoordinator.identityCapability,
-      case .unavailable =
-      identityPersistenceOperationCoordinator
-        .sharedSessionCapability,
-        options.sharedSessionSync != nil
-    else {
-      return nil
-    }
-    return SharedSessionLocalMutation(
-      baseGeneration: sharedSessionDegradationBaseGeneration
-    )
-  }
-
   /// Dispatches Clerk state changes to optional internal observers.
   var internalStateChanges = ClerkInternalStateChangeEmitter()
 
@@ -389,208 +333,6 @@ public final class Clerk {
 }
 
 extension Clerk {
-  /// Internal helper method that performs the actual configuration work.
-  @MainActor
-  func performConfiguration(
-    publishableKey: String,
-    options: Clerk.Options,
-    dependencyFactory: ConfigurationDependencyFactory? = nil
-  ) throws {
-    func makeDependencies(
-      forceVolatileIdentityPersistence: PersistenceFailureKind? = nil
-    ) throws -> DependencyContainer {
-      if let dependencyFactory {
-        return try dependencyFactory(
-          publishableKey,
-          options,
-          forceVolatileIdentityPersistence
-        )
-      }
-      return try DependencyContainer(
-        publishableKey: publishableKey,
-        options: options,
-        runtimeScope: runtimeScope,
-        deferSharedSessionAdoption: true,
-        persistenceFailureBehavior: .useVolatileStorage,
-        forceVolatileIdentityPersistence: forceVolatileIdentityPersistence
-      )
-    }
-
-    var dependencies = try makeDependencies()
-    switch recoverAppContainerIdentityClearBeforeBootstrap(
-      protecting: dependencies
-    ) {
-    case .notNeeded:
-      break
-    case .recovered:
-      // Recovery can establish the durable adoption barrier after the first
-      // container selected its storage mode. Rebuild only on this exceptional
-      // path so the installed runtime uses the recovered topology.
-      dependencies = try makeDependencies()
-    case .blocked(
-      let blockReason,
-      let recoveryFailure,
-      let requiresAtomicRouting
-    ):
-      // A partially completed recovery can also establish the adoption
-      // barrier before a later cleanup step fails. Rebuild before installing
-      // the blocked runtime so a foreground retry cannot resume with stale
-      // persistence routing.
-      let rebuiltDependencies = try makeDependencies()
-      if requiresAtomicRouting,
-         !rebuiltDependencies.usesVolatileIdentityPersistence,
-         rebuiltDependencies.atomicIdentityStore == nil
-      {
-        dependencies = try makeDependencies(
-          forceVolatileIdentityPersistence: recoveryFailure
-        )
-      } else {
-        dependencies = rebuiltDependencies
-      }
-      installConfiguration(
-        dependencies: dependencies,
-        bootstrapBlockReason: blockReason
-      )
-      return
-    }
-
-    do {
-      try SharedSessionOwnerSlotClearRecovery.recoverIfNeeded(
-        in: dependencies.sharedSessionOwnerSlotClearRecovery
-      )
-    } catch {
-      let failure = PersistenceFailureKind.classify(error)
-      guard failure.permitsVolatileIdentityFallback else {
-        installConfiguration(
-          dependencies: dependencies,
-          bootstrapBlockReason: failure.blockReason
-        )
-        return
-      }
-      dependencies = try DependencyContainer(
-        publishableKey: publishableKey,
-        options: options,
-        runtimeScope: runtimeScope,
-        deferSharedSessionAdoption: true,
-        forceVolatileIdentityPersistence: failure
-      )
-      ClerkLogger.logError(
-        error,
-        message: "Durable Clerk clear recovery is unavailable. Continuing with isolated in-memory identity for this process."
-      )
-    }
-
-    if let bootstrapFailure =
-      dependencies.identityPersistenceBootstrapFailure
-    {
-      installConfiguration(
-        dependencies: dependencies,
-        bootstrapBlockReason: bootstrapFailure.blockReason
-      )
-      return
-    }
-
-    do {
-      try dependencies.probeLocalIdentityPersistence()
-      try dependencies.discardPendingPublicationWhenSharedSyncDisabled()
-    } catch {
-      let failure = PersistenceFailureKind.classify(error)
-      guard failure.permitsVolatileIdentityFallback else {
-        installConfiguration(
-          dependencies: dependencies,
-          bootstrapBlockReason: failure.blockReason
-        )
-        return
-      }
-      dependencies = try DependencyContainer(
-        publishableKey: publishableKey,
-        options: options,
-        runtimeScope: runtimeScope,
-        deferSharedSessionAdoption: true,
-        forceVolatileIdentityPersistence: failure
-      )
-    }
-
-    installConfiguration(dependencies: dependencies)
-  }
-
-  /// Internal helper method that installs a prebuilt dependency container and starts managers.
-  @MainActor
-  func performConfiguration(dependencies: any Dependencies) throws {
-    switch recoverAppContainerIdentityClearBeforeBootstrap(
-      protecting: dependencies
-    ) {
-    case .notNeeded, .recovered:
-      break
-    case .blocked(let blockReason, _, _):
-      installConfiguration(
-        dependencies: dependencies,
-        bootstrapBlockReason: blockReason
-      )
-      return
-    }
-    try SharedSessionOwnerSlotClearRecovery.recoverIfNeeded(
-      in: dependencies.sharedSessionOwnerSlotClearRecovery
-    )
-    try dependencies.probeLocalIdentityPersistence()
-    installConfiguration(dependencies: dependencies)
-  }
-
-  /// Completes an explicit clear before any persisted identity can be hydrated.
-  /// A valid unresolved intent always blocks identity; an unreadable or malformed
-  /// intent fails closed without guessing at a deletion target.
-  private func recoverAppContainerIdentityClearBeforeBootstrap(
-    protecting currentDependencies: any Dependencies
-  )
-    -> AppContainerIdentityClearBootstrapResult
-  {
-    let intent: AppContainerIdentityClearIntent
-    do {
-      guard let pendingIntent = try appContainerIdentityClearIntentStore.load()
-      else {
-        return .notNeeded
-      }
-      intent = pendingIntent
-    } catch {
-      let failure = PersistenceFailureKind.classify(error)
-      ClerkLogger.logError(
-        error,
-        message: "Clerk could not read its pending app-container identity clear."
-      )
-      return .blocked(
-        reason: failure.blockReason,
-        recoveryFailure: failure,
-        requiresAtomicRouting: false
-      )
-    }
-
-    do {
-      try recoverAppContainerIdentityClear(
-        intent,
-        protecting: currentDependencies
-      )
-      try appContainerIdentityClearIntentStore.remove(
-        matching: intent.transactionID
-      )
-      return .recovered
-    } catch {
-      let failure = PersistenceFailureKind.classify(error)
-      ClerkLogger.logError(
-        error,
-        message: "Clerk could not complete its pending durable identity clear."
-      )
-      return .blocked(
-        reason: .pendingClear,
-        recoveryFailure: failure,
-        requiresAtomicRouting:
-        appContainerIdentityClearCanChangeCurrentPersistenceRouting(
-          intent,
-          protecting: currentDependencies
-        )
-      )
-    }
-  }
-
   private func recoverAppContainerIdentityClearStrictlyIfNeeded(
     protecting currentDependencies: any Dependencies
   ) throws -> Bool {
@@ -605,32 +347,6 @@ extension Clerk {
       matching: intent.transactionID
     )
     return true
-  }
-
-  func appContainerIdentityClearCanChangeCurrentPersistenceRouting(
-    _ intent: AppContainerIdentityClearIntent,
-    protecting currentDependencies: any Dependencies
-  ) -> Bool {
-    guard intent.ownerSlot != nil else { return false }
-    let currentConfiguration =
-      currentDependencies.configurationManager
-    let currentFingerprint = SharedSessionNamespace(
-      frontendApiUrl: currentConfiguration.frontendApiUrl,
-      publishableKey: currentConfiguration.publishableKey
-    ).fingerprint
-    guard currentFingerprint == intent.instanceFingerprint,
-          let currentIntent =
-          try? Self.makeAppContainerIdentityClearIntent(
-            dependencies: currentDependencies,
-            options: currentConfiguration.options,
-            frontendApiUrl: currentConfiguration.frontendApiUrl,
-            publishableKey: currentConfiguration.publishableKey
-          )
-    else {
-      return false
-    }
-    return currentIntent.ownerIdentifier == intent.ownerIdentifier
-      && currentIntent.stableIdentity == intent.stableIdentity
   }
 
   /// Installs dependencies whose pending clear recovery has already been checked.
@@ -753,7 +469,7 @@ extension Clerk {
     }
   }
 
-  private func installWatchConnectivityIfAvailable() {
+  func installWatchConnectivityIfAvailable() {
     guard options.watchConnectivityEnabled,
           identityPersistenceOperationCoordinator.isIdentityReady,
           case .durable =
@@ -814,195 +530,6 @@ extension Clerk {
         ClerkLogger.logError(error, message: "Failed to load client")
       }
     }
-  }
-
-  @MainActor
-  private func startSharedSessionSyncIfNeeded(
-    dependencies: any Dependencies,
-    cacheManager: CacheManager,
-    bootstrapOwnership: PersistenceTransitionOwnership
-  ) -> Task<Bool, Never>? {
-    guard !dependencies.usesVolatileIdentityPersistence,
-          options.sharedSessionSync != nil
-    else {
-      return nil
-    }
-    guard options.keychainConfig.normalizedAccessGroup != nil,
-          let ownerIdentifier = dependencies.sharedSessionOwnerIdentifier,
-          !ownerIdentifier.isEmpty,
-          let localIdentityStore = dependencies.atomicIdentityStore
-    else {
-      settleUnavailableSharedSession(
-        error: ClerkClientError(
-          message: "Shared session sync requires a Keychain access group, bundle identifier, and app-local identity store."
-        ),
-        coordinator: nil,
-        cacheManager: cacheManager,
-        ownership: bootstrapOwnership
-      )
-      return nil
-    }
-
-    do {
-      let didAdoptSharedSession =
-        try (dependencies as? DependencyContainer)?
-          .performDeferredSharedSessionAdoptionIfNeeded() ?? false
-      if didAdoptSharedSession {
-        cacheManager.loadProvisionalLegacyClientForPresentation()
-      }
-      let localMutationPreparation =
-        try SharedSessionRecoveryReconciler
-          .preparePendingLocalMutationForActivation(
-            in: dependencies
-          )
-      if localMutationPreparation?.publishedDestinationSlot != nil {
-        Self.notifySharedSessionTopologyChange(in: dependencies)
-      }
-      try identityPersistenceOperationCoordinator.validate(
-        bootstrapOwnership,
-        operation: .bootstrap
-      )
-      let coordinator = try makeSharedSessionSyncCoordinator(
-        dependencies: dependencies,
-        ownerIdentifier: ownerIdentifier,
-        localIdentityStore: localIdentityStore
-      )
-      sharedSessionSyncCoordinator = coordinator
-      internalStateChanges.addObserver(coordinator)
-      let initialTask = coordinator.start()
-      return Task { @MainActor [weak self, weak coordinator] in
-        _ = await initialTask.value
-        guard let self, let coordinator else { return false }
-        do {
-          try identityPersistenceOperationCoordinator.validate(
-            bootstrapOwnership,
-            operation: .bootstrap
-          )
-          try await coordinator.waitForInitialReconciliation()
-          try identityPersistenceOperationCoordinator.validate(
-            bootstrapOwnership,
-            operation: .bootstrap
-          )
-          identityPersistenceOperationCoordinator.setSharedSessionCapability(
-            .active
-          )
-          sharedSessionDegradationBaseGeneration = nil
-          identityPersistenceOperationCoordinator.finish(bootstrapOwnership)
-          installWatchConnectivityIfAvailable()
-          return true
-        } catch is CancellationError {
-          return false
-        } catch {
-          settleUnavailableSharedSession(
-            error: error,
-            failure:
-            coordinator.lastReconciliationFailureKind,
-            coordinator: coordinator,
-            cacheManager: cacheManager,
-            ownership: bootstrapOwnership
-          )
-          return false
-        }
-      }
-    } catch {
-      settleUnavailableSharedSession(
-        error: error,
-        coordinator: nil,
-        cacheManager: cacheManager,
-        ownership: bootstrapOwnership
-      )
-      return nil
-    }
-  }
-
-  private func makeSharedSessionSyncCoordinator(
-    dependencies: any Dependencies,
-    ownerIdentifier: String,
-    localIdentityStore: any SharedSessionLocalIdentityStoring
-  ) throws -> SharedSessionSyncCoordinator {
-    let namespace = SharedSessionNamespace(
-      frontendApiUrl: frontendApiUrl,
-      publishableKey: publishableKey
-    )
-    let slotStore = try SharedSessionOwnerSlotStore(
-      keychainConfig: options.keychainConfig,
-      namespace: namespace,
-      ownerIdentifier: ownerIdentifier
-    )
-    return SharedSessionSyncCoordinator(
-      ownerIdentifier: ownerIdentifier,
-      instanceFingerprint: namespace.fingerprint,
-      slotStore: slotStore,
-      localIdentityStore: localIdentityStore,
-      localIdentityIO: dependencies.atomicIdentityIO,
-      notifier: SharedSessionSyncDarwinNotifier(
-        keychainConfig: options.keychainConfig,
-        instanceFingerprint: namespace.fingerprint
-      ),
-      configurationEpoch: configurationEpoch,
-      clerk: self
-    )
-  }
-
-  private func settleUnavailableSharedSession(
-    error: any Error,
-    failure: PersistenceFailureKind? = nil,
-    coordinator: SharedSessionSyncCoordinator?,
-    cacheManager: CacheManager,
-    ownership: PersistenceTransitionOwnership
-  ) {
-    guard identityPersistenceOperationCoordinator.isActive(
-      ownership,
-      operation: .bootstrap
-    )
-    else {
-      return
-    }
-    ClerkLogger.logError(
-      error,
-      message: "Shared session sync is unavailable. Clerk will continue with durable app-local identity."
-    )
-    coordinator?.deactivate()
-    if let observedGeneration = coordinator?.currentMaximumGeneration,
-       observedGeneration > 0
-    {
-      sharedSessionDegradationBaseGeneration = max(
-        sharedSessionDegradationBaseGeneration ?? 0,
-        observedGeneration
-      )
-    }
-    if sharedSessionSyncCoordinator === coordinator {
-      sharedSessionSyncCoordinator = nil
-    }
-    if let coordinator {
-      internalStateChanges.removeObserver(coordinator)
-    }
-    do {
-      try dependencies.atomicIdentityStore?
-        .convertPendingPublicationToUnpublishedLocalMutation(
-          baseGeneration: sharedSessionDegradationBaseGeneration
-        )
-    } catch {
-      ClerkLogger.logError(
-        error,
-        message: "Shared session sync failed while preserving the pending app-local identity."
-      )
-      identityController.resetRuntimeIdentity()
-      identityPersistenceOperationCoordinator.block(
-        ownership,
-        reason: PersistenceFailureKind.classify(error).blockReason
-      )
-      return
-    }
-    identityController.resetRuntimeIdentity()
-    cacheManager.loadCachedData()
-    identityPersistenceOperationCoordinator.setSharedSessionCapability(
-      .unavailable(
-        failure ?? PersistenceFailureKind.classify(error)
-      )
-    )
-    identityPersistenceOperationCoordinator.finish(ownership)
-    installWatchConnectivityIfAvailable()
   }
 
   /// Configures the shared Clerk instance.
@@ -1139,12 +666,9 @@ extension Clerk {
     defer { endRuntimeReconfiguration() }
 
     if let existing = _shared {
-      let preparation = try await existing.prepareReconfiguration(
+      return try await existing.performReconfiguration(
         publishableKey: publishableKey,
         options: options
-      )
-      return try await existing.installPreparedReconfiguration(
-        preparation
       )
     }
 
@@ -1267,21 +791,6 @@ extension Clerk {
 
     try await auth.handle(route)
     return true
-  }
-
-  @MainActor
-  private func resetRuntimeStateForReconfiguration() async {
-    await SessionTokenFetcher.shared.reset()
-    await SessionTokensCache.shared.clear()
-
-    identityController.resetRuntimeIdentity()
-    environment = nil
-    sessionsByUserId = [:]
-    WebAuthentication.cancelCurrentSession()
-
-    #if canImport(AuthenticationServices) && !os(watchOS)
-    PasskeyHelper.cancelCurrentAuthorization()
-    #endif
   }
 }
 
@@ -1507,310 +1016,6 @@ extension Clerk {
     return clerk
   }
 
-  private func prepareReconfiguration(
-    publishableKey: String,
-    options: Clerk.Options
-  ) async throws -> PreparedReconfiguration {
-    let previousReadiness = persistenceStatus.readiness
-    identityPersistenceOperationCoordinator.cancelActiveTransition()
-    let ownership = try identityPersistenceOperationCoordinator
-      .beginReconfiguration(epoch: configurationEpoch)
-    suspendWatchConnectivityForPersistenceTransition()
-    let nextEpoch = nextConfigurationEpoch
-
-    do {
-      let target = try DependencyContainer(
-        publishableKey: publishableKey,
-        options: options,
-        runtimeScope: .init(
-          epoch: nextEpoch,
-          runtimeState: runtimeState
-        ),
-        deferSharedSessionAdoption: true
-      )
-      try prepareReconfigurationTarget(target)
-      let plan = reconfigurationPlan(with: target)
-      if plan == .preserveIdentityAndMigrateSlot {
-        try await settlePendingPublicationForTopologyChange()
-      }
-      try identityPersistenceOperationCoordinator.validate(
-        ownership,
-        operation: .reconfigure
-      )
-      return try PreparedReconfiguration(
-        ownership: ownership,
-        nextEpoch: nextEpoch,
-        dependencies: target,
-        plan: plan,
-        rollbackState: captureReconfigurationRollbackState(),
-        volatileIdentity: volatileIdentityForReconfiguration()
-      )
-    } catch {
-      restoreAfterFailedReconfigurationPreparation(
-        ownership: ownership,
-        previousReadiness: previousReadiness
-      )
-      throw error
-    }
-  }
-
-  private func prepareReconfigurationTarget(
-    _ target: DependencyContainer
-  ) throws {
-    try SharedSessionOwnerSlotClearRecovery.recoverIfNeeded(
-      in: target.sharedSessionOwnerSlotClearRecovery
-    )
-    try target.probeLocalIdentityPersistence()
-    try target.discardPendingPublicationWhenSharedSyncDisabled()
-  }
-
-  private func volatileIdentityForReconfiguration() throws
-    -> ClerkIdentitySnapshot?
-  {
-    guard dependencies.usesVolatileIdentityPersistence else {
-      return nil
-    }
-    return try identityController.currentIdentitySnapshot()
-  }
-
-  private func restoreAfterFailedReconfigurationPreparation(
-    ownership: PersistenceTransitionOwnership,
-    previousReadiness: PersistenceStatus.Readiness
-  ) {
-    if case .blocked(let reason) = previousReadiness {
-      identityPersistenceOperationCoordinator.block(
-        ownership,
-        reason: reason
-      )
-    } else {
-      identityPersistenceOperationCoordinator.finish(ownership)
-    }
-    resumeWatchConnectivityAfterPersistenceTransition()
-  }
-
-  private func installPreparedReconfiguration(
-    _ preparation: PreparedReconfiguration
-  ) async throws -> Clerk {
-    setConfigurationEpoch(to: preparation.nextEpoch)
-    await cleanupManagersAndDrainCache(
-      deleteSharedSessionOwnerSlot: false,
-      cancelPersistenceTransition: false
-    )
-
-    var topologyRollback: SharedSessionTopologyMigration.Rollback?
-    do {
-      try identityPersistenceOperationCoordinator.validate(
-        preparation.ownership,
-        operation: .reconfigure,
-        expectedEpoch: preparation.nextEpoch
-      )
-      topologyRollback = try await prepareIdentityForReconfiguration(
-        preparation
-      )
-      try await retireReconfigurationSource(
-        preparation,
-        topologyRollback: topologyRollback
-      )
-    } catch {
-      restoreTopologyMigration(topologyRollback)
-      restoreAfterFailedReconfiguration(preparation.rollbackState)
-      throw error
-    }
-
-    await resetRuntimeStateForReconfiguration()
-    installConfiguration(dependencies: preparation.dependencies)
-    return self
-  }
-
-  private func prepareIdentityForReconfiguration(
-    _ preparation: PreparedReconfiguration
-  ) async throws -> SharedSessionTopologyMigration.Rollback? {
-    switch preparation.plan {
-    case .preserveIdentityAndSlot:
-      guard let identity = preparation.volatileIdentity else { return nil }
-      return try prepareLocalMutationForReconfiguration(
-        identity,
-        baseGeneration: sharedSessionDegradationBaseGeneration,
-        dependencies: preparation.dependencies
-      )
-    case .preserveIdentityAndMigrateSlot:
-      return try prepareMigratedIdentityForReconfiguration(preparation)
-    case .replaceIdentity:
-      try await Self.clearLocalClerkStorageStrictly(
-        in: preparation.dependencies
-      )
-      try await Self.clearLocalClerkStorageStrictly(
-        in: preparation.rollbackState.dependencies,
-        deleteSharedSessionOwnerSlot: false
-      )
-      try preparation.dependencies
-        .markSharedSessionAdoptedWithoutMigratingCredentialsIfNeeded()
-      return nil
-    }
-  }
-
-  private func prepareMigratedIdentityForReconfiguration(
-    _ preparation: PreparedReconfiguration
-  ) throws -> SharedSessionTopologyMigration.Rollback? {
-    if let identity = preparation.volatileIdentity {
-      return try prepareLocalMutationForReconfiguration(
-        identity,
-        baseGeneration: sharedSessionDegradationBaseGeneration,
-        dependencies: preparation.dependencies
-      )
-    }
-    if let sourceRecord = try preparation.rollbackState
-      .dependencies.atomicIdentityStore?.loadRecord(),
-      sourceRecord.hasUnpublishedLocalMutation,
-      let identity = sourceRecord.acceptedIdentity
-    {
-      return try prepareLocalMutationForReconfiguration(
-        identity,
-        baseGeneration: sourceRecord.sharedSessionBaseGeneration,
-        dependencies: preparation.dependencies
-      )
-    }
-    return try Self.prepareAcceptedIdentityForTopologyChange(
-      from: preparation.rollbackState.dependencies,
-      to: preparation.dependencies
-    )
-  }
-
-  private func prepareLocalMutationForReconfiguration(
-    _ identity: ClerkIdentitySnapshot,
-    baseGeneration: UInt64?,
-    dependencies: any Dependencies
-  ) throws -> SharedSessionTopologyMigration.Rollback? {
-    try SharedSessionRecoveryReconciler
-      .prepareLocalMutationForActivation(
-        identity: identity,
-        baseGeneration: baseGeneration,
-        in: dependencies
-      )
-  }
-
-  private func retireReconfigurationSource(
-    _ preparation: PreparedReconfiguration,
-    topologyRollback: SharedSessionTopologyMigration.Rollback?
-  ) async throws {
-    if !preparation.plan.reusesOwnerSlot {
-      try await SharedSessionOwnerSlotCleanup.deleteIfConfigured(
-        in: preparation.rollbackState.dependencies
-      )
-    }
-    if topologyRollback?.publishedDestinationSlot != nil {
-      Self.notifySharedSessionTopologyChange(
-        in: preparation.dependencies
-      )
-    }
-    try identityPersistenceOperationCoordinator.validate(
-      preparation.ownership,
-      operation: .reconfigure,
-      expectedEpoch: preparation.nextEpoch
-    )
-  }
-
-  private func restoreTopologyMigration(
-    _ rollback: SharedSessionTopologyMigration.Rollback?
-  ) {
-    guard let rollback else { return }
-    do {
-      try rollback.restore()
-    } catch {
-      ClerkLogger.logError(
-        error,
-        message: "Failed to roll back shared-session topology migration"
-      )
-    }
-  }
-
-  private func captureReconfigurationRollbackState() -> ReconfigurationRollbackState {
-    ReconfigurationRollbackState(
-      configurationEpoch: configurationEpoch,
-      dependencies: dependencies,
-      identity: identityController.captureRollbackState()
-    )
-  }
-
-  private func restoreAfterFailedReconfiguration(
-    _ state: ReconfigurationRollbackState
-  ) {
-    setConfigurationEpoch(to: state.configurationEpoch)
-    identityController.restoreRollbackState(state.identity)
-    installConfiguration(dependencies: state.dependencies)
-  }
-
-  private func settlePendingPublicationForTopologyChange() async throws {
-    try await sharedSessionSyncCoordinator?
-      .settlePendingPublicationForTopologyChange()
-
-    guard try dependencies.atomicIdentityStore?
-      .loadPendingPublication() == nil
-    else {
-      throw SharedSessionSyncCoordinatorError.pendingPublicationDidNotSettle
-    }
-  }
-
-  private static func prepareAcceptedIdentityForTopologyChange(
-    from source: any Dependencies,
-    to destination: any Dependencies
-  ) throws -> SharedSessionTopologyMigration.Rollback? {
-    let sourceRecord = try source.atomicIdentityStore?.loadRecord()
-    guard sourceRecord?.pendingPublication == nil else {
-      throw SharedSessionSyncCoordinatorError.pendingPublicationDidNotSettle
-    }
-    guard let identity = sourceRecord?.acceptedIdentity,
-          let destinationIdentityStore = destination.atomicIdentityStore
-    else {
-      return nil
-    }
-
-    let sourceTopology = SharedSessionSlotTopology(dependencies: source)
-    let destinationTopology = SharedSessionSlotTopology(dependencies: destination)
-    let sourceOwnerToExclude: String? = if let sourceTopology,
-                                           let destinationTopology,
-                                           sourceTopology.hasSameStore(as: destinationTopology)
-    {
-      sourceTopology.ownerIdentifier
-    } else {
-      nil
-    }
-
-    return try SharedSessionTopologyMigration.prepare(
-      identity: identity,
-      destinationIdentityStore: destinationIdentityStore,
-      destinationSlotStore: destinationTopology?.makeOwnerSlotStore(),
-      destinationInstanceFingerprint: destinationTopology?.instanceFingerprint,
-      destinationOwnerIdentifier: destinationTopology?.ownerIdentifier,
-      excludingSourceOwnerIdentifier: sourceOwnerToExclude
-    )
-  }
-
-  static func notifySharedSessionTopologyChange(
-    in dependencies: any Dependencies
-  ) {
-    guard let topology = SharedSessionSlotTopology(dependencies: dependencies) else { return }
-    SharedSessionSyncDarwinNotifier(
-      keychainConfig: topology.keychainConfig,
-      instanceFingerprint: topology.instanceFingerprint
-    ).post()
-  }
-
-  private func reconfigurationPlan(
-    with newDependencies: any Dependencies
-  ) -> ReconfigurationPlan {
-    let preservesIdentity = publishableKey
-      == newDependencies.configurationManager.publishableKey
-      && (options.sharedSessionSync != nil
-        || newDependencies.configurationManager.options.sharedSessionSync != nil)
-    guard preservesIdentity else {
-      return .replaceIdentity
-    }
-    return canReuseSharedSessionOwnerSlot(with: newDependencies)
-      ? .preserveIdentityAndSlot
-      : .preserveIdentityAndMigrateSlot
-  }
-
   func isCurrentConfigurationEpoch(_ epoch: ClerkConfigurationEpoch) -> Bool {
     configurationEpoch == epoch
   }
@@ -1890,20 +1095,6 @@ extension Clerk {
     await cacheManager?.shutdownAndDrain()
     cacheManager = nil
     teardownNonCacheManagers()
-  }
-
-  func canReuseSharedSessionOwnerSlot(
-    with newDependencies: any Dependencies
-  ) -> Bool {
-    guard let currentTopology = SharedSessionSlotTopology(
-      dependencies: dependencies
-    ),
-      let newTopology = SharedSessionSlotTopology(dependencies: newDependencies)
-    else {
-      return false
-    }
-
-    return currentTopology.hasSameOwnerSlot(as: newTopology)
   }
 
   private func resetManagerStateForCleanup(finishAuthEventStreams: Bool) {
